+++ public/analysis.js
@@ -1,0 +1,6 @@
+onmessage = (evt) => {
+    const channel = evt.data.channel
+    const payload = evt.data.payload
+
+    console.log(`${channel}: ${payload}`)
+}
