+++ src/models/WorkerMessage.js
@@ -1,0 +1,8 @@
+class WorkerMessage {
+    constructor (channel, payload) {
+        this.channel = channel  // String
+        this.payload = payload  // Whatever
+    }
+}
+
+export default WorkerMessage
